feat: simplify usage of sixFormUtil directive
