+++ examples/angular/src/app/pages/form/form.component.ts
@@ -1,7 +1,7 @@
-import { ChangeDetectionStrategy, Component, OnInit, ViewChild } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { AbstractControl, FormBuilder, FormGroup, ValidationErrors, Validators } from '@angular/forms';
 import { BehaviorSubject } from 'rxjs';
-import { SixFormUtilDirective, SixUiLibraryValidators } from '@six-group/ui-library-angular';
+import { SixUiLibraryValidators } from '@six-group/ui-library-angular';
 
 type UserGroup = 'admin' | 'developer' | 'user';
 type Status = 'enabled' | 'disabled' | 'temporary';
@@ -13,8 +13,7 @@ type Interest = 'sport' | 'music' | 'movies';
   styleUrls: ['./form.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class FormComponent implements OnInit {
-  @ViewChild(SixFormUtilDirective) sixFormUtil!: SixFormUtilDirective;
+export class FormComponent {
   minDate = addDays(removeTime(new Date()), 1);
   maxDate = addDays(removeTime(new Date()), -1);
 
@@ -69,15 +68,11 @@ export class FormComponent implements OnInit {
     });
   }
 
-  onSubmit() {
-    if (!this.userForm.valid) {
-      this.sixFormUtil.focusInvalidField();
-    }
+  onSubmit(event: SubmitEvent) {
+    console.log('form submitted', event);
     console.log('valid:', this.userForm.valid);
     console.log('form data:', this.userForm.value);
   }
-
-  ngOnInit(): void {}
 }
 
 function getFormDebug(form: FormGroup) {
